feat(app-switcher): separate menu from button impl
