+++ src/lib/app-switcher/app-switcher-menu/app-switcher-menu.component.html
@@ -1,0 +1,30 @@
+<td-menu class="app-switcher-menu">
+  <!--header-->
+  <div td-menu-header class="app-switcher-header">
+    <mat-icon
+      class="logo-icon"
+      [svgIcon]="_themeService.map({ 'dark-theme': 'td-logo:on-dark', 'light-theme': 'td-logo:on-light' }) | async"
+    ></mat-icon>
+    <div class="app-switcher-header-title mat-body-2">{{ 'APP_SWITCHER.HEADER' | translate }}</div>
+  </div>
+  <!--content-->
+  <ng-container *ngFor="let product of productList; last as isLast">
+    <a mat-menu-item [href]="product.href">
+      <mat-icon
+        [class]="product.iconClasses"
+        *ngIf="product.svgIcon; else iconTemplate"
+        [svgIcon]="product.svgIcon"
+      ></mat-icon>
+      <ng-template #iconTemplate>
+        <mat-icon [class]="product.iconClasses">{{ product.icon || 'apps' }}</mat-icon>
+      </ng-template>
+      {{ product.text }}
+    </a>
+    <mat-divider *ngIf="product.divider && !isLast"></mat-divider>
+  </ng-container>
+  <!--footer-->
+  <a mat-menu-item td-menu-footer class="explore-more" [href]="exploreMoreLink">
+    <span class="mat-caption">{{ 'APP_SWITCHER.EXPLORE_MORE' | translate }}</span>
+    <mat-icon>launch</mat-icon>
+  </a>
+</td-menu>
