--- src/lib/app-switcher/app-switcher.component.scss
@@ -1,22 +1,0 @@
-.app-switcher-menu {
-  min-width: 250px;
-
-  .app-switcher-header[td-menu-header] {
-    padding: 24px 0;
-    .logo-icon {
-      width: 100px;
-      margin-bottom: 8px;
-    }
-  }
-
-  .explore-more {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-
-    .mat-icon {
-      font-size: 16px;
-      margin-right: 0;
-    }
-  }
-}
