+++ src/lib/app-switcher/public_api.ts
@@ -1,2 +1,3 @@
 export * from './app-switcher.module';
 export * from './app-switcher.component';
+export * from './app-switcher-menu/app-switcher-menu.component';
